re-structure

+++ src/plugins/writing/remove_from_list.py
@@ -1,19 +1,17 @@
 import re
 import time
-
 from hydrogram import Client, filters
 from hydrogram.enums import ChatAction
+from database import NotesDB
 
-from db import NotesDB
 
-
-@Client.on_message(filters.command("add"))
-def add_url(c, m):
+@Client.on_message(filters.command("delete"))
+def delete_url(c, m):
     notes = NotesDB()
     m.reply_chat_action(ChatAction.TYPING)
     user_id = m.from_user.id
     filename = f"{user_id}"
-    if m.from_user.id == 5665225938:
+    if user_id == 5665225938:
         filename = "v2ray"
     text = m.text
     if m.reply_to_message:
@@ -28,27 +26,25 @@ def add_url(c, m):
         c.delete_messages(m.chat.id, err.id)
         m.delete()
         return
-    li = 0
+    worked = None
     for url in urls:
+        worked = False
         try:
-            notes.add(filename, url)
-            li += 1
+            notes.remove(filename, url)
+            worked = True
         except Exception as e:
-            print(e)
-            pass
-    if li != len(urls):
-        x = len(urls) - li
-        temp = m.reply(f"{x} URL trùng lặp sẽ không được thêm lại")
-        time.sleep(5)
-        c.delete_messages(m.chat.id, temp.id)
-    done = m.reply_text(f"Đã thêm {li} URL")
-    time.sleep(10)
-    c.delete_messages(m.chat.id, done.id)
+            err = m.reply_text(f"Error: {e}")
+            time.sleep(10)
+            c.delete_messages(m.chat.id, err.id)
+    if worked:
+        done = m.reply_text(f"Đã xoá {len(urls)} URL")
+        time.sleep(10)
+        c.delete_messages(m.chat.id, done.id)
     m.delete()
 
 
-@Client.on_message(filters.command("share"))
-def share_url(c, m):
+@Client.on_message(filters.command("deletesharelist"))
+def delete_share_url(c, m):
     notes = NotesDB()
     m.reply_chat_action(ChatAction.TYPING)
     text = m.text
@@ -62,20 +58,20 @@ def share_url(c, m):
         err = m.reply_text("Vui lòng cung cấp URL")
         time.sleep(10)
         c.delete_messages(m.chat.id, err.id)
+        m.delete()
         return
-    li = 0
+    worked = None
     for url in urls:
+        worked = False
         try:
-            notes.add("share", url)
-            li += 1
+            notes.remove("share", url)
+            worked = True
         except Exception as e:
-            print(e)
-            pass
-    if li != len(urls):
-        x = len(urls) - li
-        temp = m.reply(f"{x} URL trùng lặp sẽ không được thêm lại")
+            err = m.reply_text(f"Error: {e}")
+            time.sleep(10)
+            c.delete_messages(m.chat.id, err.id)
+    if worked:
+        done = m.reply_text(f"Đã xoá {len(urls)} URL")
         time.sleep(10)
-        c.delete_messages(m.chat.id, temp.id)
-    done = m.reply_text(f"Đã thêm {li} URL")
-    time.sleep(10)
-    done.delete()
+        c.delete_messages(m.chat.id, done.id)
+    m.delete()
